외부 파일 로드하는 기능 별도로 분리 분리 후 빌드 안되는 것들 정리 및 코드 정리

+++ include/component/VKSkymapModelDescriptor.h
@@ -33,7 +33,7 @@ struct VKSkyMapModelDescriptor : public VKDescriptor {
   void setObject(object::Object* object) { this->objects.push_back(object); }
 
   virtual uint16_t getDescriptorCount() {
-    return this->objects.size();
+      return static_cast<uint16_t>(this->objects.size());
   }  // 디스크립터 개수 반환
 
   virtual VkPipelineLayout getPipelineLayout() {
